Restore TAG_CHANGES in BOARD_VISUAL_STATE=2 TRIGGER blocks (#16)
There's some logic in the filter to strip out tag changes that share a
TRIGGER block with a tag change that sets BOARD_VISUAL_STATE to 2. We
believe that for some time now, BOARD_VISUAL_STATEs 1 or 2 are related to
shop manipulation code and not part of what we want to capture in combat.

However! In a recent BGS log, there are tag changes that seem to represent
enchantments (i.e., ATK and HEALTH buffs) being applied to entities in a
BOARD_VISUAL_STATE=2 block; whereas in older BGS logs, that happens in
a separate block.

This change removes the logic that was stripping all BOARD_VISUAL_STATE=2
block tag changes.